# Faster Element Symbol Handling