- Replaces lookup of const atomic symbols with the existing constant definitions.
- Add some extra EL_NUMBER_ constants
- Optimise the lookup for common cases
- Replace lazy initialisation of some element arrays with a much simpler and less error prone declaration.

Testing on ChEMBL I see 2.4 million identical InChI's and runs ~9 seconds faster (M1 Pro).
Overall I mainly was wincing and some of the lookups inside nested ternaries (e.g. ichitaut.c)

```
main:
INCHI_EXE/bin/Linux/inchi-1 /data/chembl_34.sdf -AuxNone -NoWarnings -NoLabel
264.73s user 24.96s system 99% cpu 4:51.59 total
PR:
INCHI_EXE/bin/Linux/inchi-1 /data/chembl_34.sdf -AuxNone -NoWarnings -NoLabel
257.91s user 23.88s system 99% cpu 4:42.53 total
```

I may have left some unused variables when I removed the lazy ``if (!en){ }`` initialisations but hopefully got them all.