@johnmay 2024-07-25
I've made some more changes to where the element arrays get setup in various places (mainly strutil.c ion_pair handling). I also found some bugs from when the thread safety was added, nothing major but some had an out of bounds checks which wouldn't have been obvious to the static analysis.