@djb-rwth 2024-07-25
Hi @johnmay,
Thank you for these useful suggestions.
These changes make sense as using constants instead of calling a function can indeed make a notable increase in performance, especially whilst processing large datasets.

The only suggestion from my side is to add these changes to the `rwth` branch instead of `main` directly as I need to add a file which should fix a rather serious bug; afterwards everything shall be merged into the `main` branch.